Revert "Initial commit (#104)" (#106)
This reverts commit fe0a557360b3baa9bb51ab1a191e32454d7602ba.

+++ OrdermentumSDK/Models/Cart.swift
@@ -19,13 +19,13 @@ public struct Cart {
 
 public struct CartProduct {
     public init() {}
-    public init(product: Product, quantity: Float, comment: String? = nil) {
+    public init(product: Product, quantity: Int, comment: String? = nil) {
         self.product = product
         self.quantity = quantity
         self.comment = comment
     }
     public var product: Product = Product()
-    public var quantity: Float = 0.0
+    public var quantity: Int = 0
     public var comment: String?
 }
